- refactoring app directory
