+++ src/Page/task-list/styles/index.tsx
@@ -1,0 +1,9 @@
+import styled from "@emotion/styled";
+
+export const Layout = styled.div`
+  margin: 0;
+  padding: 0;
+  display: flex;
+  width: 100%;
+  height: 100%;
+`
