--- src/app/App.tsx
@@ -1,11 +1,0 @@
-import * as React from "react";
-import { MainPage } from "../Page";
-
-const App = () => {
-
-  return (
-    <MainPage/>
-  );
-}
-
-export default App;
