+++ src/app/index.tsx
@@ -1,12 +1,12 @@
-import 'tslib';
-import React from 'react';
-import ReactDOM from 'react-dom';
-import App from './App';
-import store from "../state";
-import { Provider } from "react-redux";
+import * as React from "react";
+import { LayoutBody } from "./styles";
+import { TaskListPage } from "../Page";
 
-ReactDOM.render(
-  <Provider store={store}>
-    <App/>
-  </Provider>,
-  document.getElementById('root'));
+export const App = () => {
+
+  return (
+    <LayoutBody>
+      <TaskListPage/>
+    </LayoutBody>
+  );
+}
