--- src/app/style.tsx
@@ -1,12 +1,0 @@
-import styled from "@emotion/styled";
-
-export const ContainerPage = styled.div`
-  body {
-    margin: 0;
-    padding: 0;
-  }
-  display: flex;
-  width: 100%;
-  height: 100%;
-  color: #fff;
-`
