+++ src/app/styles/index.tsx
@@ -1,0 +1,9 @@
+import styled from "@emotion/styled";
+
+export const LayoutBody = styled.body`
+  margin: 0;
+  padding: 0;
+  display: flex;
+  width: 100%;
+  height: 100%;
+`
